feat(img2img): VAEEncode + init_image 支持，修复 float FFI ABI
- C++ 适配器：新增 sd_pipeline_set_init_image（OpenCV 加载参考图），generate() 使用
  init_image + strength，未指定宽高时用参考图尺寸
- StaticPy：LatentImage 增加 image_path；新增 VAEEncode 节点；KSampler 读 denoise
  并调用 set_init_image
- 修复：TYPE_MAP float→C float（上游默认 double 导致所有 float FFI 参数 ABI 错位，
  cfg/strength 等全变垃圾值）
- 验证：img2img denoise=0.6 出图，strength 正确传递

+++ staticpy/static_translate.py
@@ -44,7 +44,9 @@ def scheme_name(name):
 # ====== C 类型映射 ======
 TYPE_MAP = {
     "int": "int",
-    "float": "double",
+    # comfycli 的 C API（sdcpp_adapter）用 32 位 float；上游默认 double 会导致
+    # foreign-procedure ABI 不匹配（float 参数变成垃圾值）。
+    "float": "float",
     "double": "double",
     "bool": "boolean",
     "str": "string",
